Disable hugeexpr1 test under JitStress modes
This test runs out of memory on x86 when its huge expressions combine with
STRESS_CLONE_EXPR.  In this case, the "OptimizationSensitive" label isn't ideal,
but the meaning is what we need: don't run this under JitStress modes.
This is disabled at build time for x86, which means that if x64-built tests are
run on other architectures, e.g. ARM or ARM64, this change won't take effect.